\name{percent_reduction}
\alias{percent_reduction}
\title{This function to calculate percent reduction year on year}
\usage{
percent_reduction(x)
}
\arguments{
  \item{x}{is a column of data. In this project,this
  function can be used to calculated the percentage
  reduction of mad cow diseased year on year}
}
\description{
This function to calculate percent reduction year on year
}
\examples{
result <- percent_reduction(x)
}

